Create README.md
Create a Simple Interest calculator using HTML and JavaScript. Take principle amount, rate of interest and time as an input in the text boxes from the user. Apply validation that amount, rate of interest and time can only be numeric. Rate of interest should not be more than 100 and time is in number of years and cannot be greater than 15 yrs. Form must be validated immediately once the text box is filled.